// csa-app/frontend/src/app/model/findings.ts
export interface Findings {
  id: number;
  run: number;
  filename: string;
  fqn: string;
  ext: string;
  line: number;
  rule: string;
  pattern: string;
  value: string;
  note: string;
  advice: string;
  level: string;
  effort: number;
  readiness: number;
  category: string;
  criticality: string;
  application: string;
  tags: string[];
  recipes: string[];
}